+++ gsplat/compression/png_compression.py
@@ -1,0 +1,338 @@
+from dataclasses import dataclass
+import os
+from typing import Any, Callable
+import json
+import torch
+from torch import Tensor
+import torch.nn.functional as F
+import numpy as np
+import imageio
+
+from gsplat.compression.sort import sort_splats
+from gsplat.utils import log_transform, inverse_log_transform
+
+
+@dataclass
+class PngCompressionStrategy:
+    """Compress splats with quantization, sorting, and K-means clustering of
+    the spherical harmonic coefficents.
+
+    Args:
+        use_sort (bool, optional): Whether to sort splats before compression. Defaults to True.
+        verbose (bool, optional): Whether to print verbose information. Default to True.
+    """
+
+    use_sort: bool = True
+    verbose: bool = True
+
+    def get_compress_fn(self, param_name: str) -> Callable:
+        compress_fn_map = {
+            "means": _compress_png_16bit,
+            "scales": _compress_png,
+            "quats": _compress_png,
+            "opacities": _compress_png,
+            "sh0": _compress_png,
+            "shN": _compress_kmeans,
+        }
+        if param_name in compress_fn_map:
+            return compress_fn_map[param_name]
+        else:
+            return _compress_npz
+
+    def get_decompress_fn(self, param_name: str) -> Callable:
+        decompress_fn_map: dict[str, Callable] = {
+            "means": _decompress_png_16bit,
+            "scales": _decompress_png,
+            "quats": _decompress_png,
+            "opacities": _decompress_png,
+            "sh0": _decompress_png,
+            "shN": _decompress_kmeans,
+        }
+        if param_name in decompress_fn_map:
+            return decompress_fn_map[param_name]
+        else:
+            return _decompress_npz
+
+    def compress(self, compress_dir: str, splats: dict[str, Tensor]) -> None:
+        # Param-specific preprocessing
+        splats["means"] = log_transform(splats["means"])
+        splats["quats"] = F.normalize(splats["quats"], dim=-1)
+
+        n_gs = len(splats["means"])
+        n_sidelen = int(n_gs**0.5)
+        n_crop = n_gs - n_sidelen**2
+        if n_crop != 0:
+            splats = self._crop_n_splats(splats, n_crop)
+            print(
+                f"Warning: Number of Gaussians was not square. Removed {n_crop} Gaussians."
+            )
+
+        if self.use_sort:
+            splats = sort_splats(splats)
+
+        meta = {}
+        for param_name in splats.keys():
+            compress_fn = self.get_compress_fn(param_name)
+            kwargs = {
+                "n_sidelen": n_sidelen,
+                "verbose": self.verbose,
+            }
+            meta[param_name] = compress_fn(
+                compress_dir, param_name, splats[param_name], **kwargs
+            )
+
+        with open(os.path.join(compress_dir, "meta.json"), "w") as f:
+            json.dump(meta, f)
+
+    def decompress(self, compress_dir: str) -> dict[str, Tensor]:
+        with open(os.path.join(compress_dir, "meta.json"), "r") as f:
+            meta = json.load(f)
+
+        splats = {}
+        for param_name, param_meta in meta.items():
+            decompress_fn = self.get_decompress_fn(param_name)
+            splats[param_name] = decompress_fn(compress_dir, param_name, param_meta)
+
+        # Param-specific postprocessing
+        splats["means"] = inverse_log_transform(splats["means"])
+        return splats
+
+    def _crop_n_splats(
+        self, splats: dict[str, Tensor], n_crop: int
+    ) -> dict[str, Tensor]:
+        opacities = splats["opacities"]
+        keep_indices = torch.argsort(opacities, descending=True)[:-n_crop]
+        for k, v in splats.items():
+            splats[k] = v[keep_indices]
+        return splats
+
+
+def _compress_png(
+    compress_dir: str, param_name: str, params: Tensor, n_sidelen: int, **kwargs
+) -> dict[str, Any]:
+    """Compress parameters with 8-bit quantization and lossless PNG compression.
+
+    Args:
+        compress_dir (str): compression directory
+        param_name (str): parameter field name
+        params (Tensor): parameters
+        n_sidelen (int): image side length
+
+    Returns:
+        dict[str, Any]: metadata
+    """
+    grid = params.reshape((n_sidelen, n_sidelen, -1))
+    mins = torch.amin(grid, dim=(0, 1))
+    maxs = torch.amax(grid, dim=(0, 1))
+    grid_norm = (grid - mins) / (maxs - mins)
+    img_norm = grid_norm.detach().cpu().numpy()
+
+    img = (img_norm * (2**8 - 1)).round().astype(np.uint8)
+    img = img.squeeze()
+    imageio.imwrite(os.path.join(compress_dir, f"{param_name}.png"), img)
+
+    meta = {
+        "shape": list(params.shape),
+        "dtype": str(params.dtype).split(".")[1],
+        "mins": mins.tolist(),
+        "maxs": maxs.tolist(),
+    }
+    return meta
+
+
+def _decompress_png(compress_dir: str, param_name: str, meta: dict[str, Any]) -> Tensor:
+    """Decompress parameters from PNG file.
+
+    Args:
+        compress_dir (str): compression directory
+        param_name (str): parameter field name
+        meta (dict[str, Any]): metadata
+
+    Returns:
+        Tensor: parameters
+    """
+    img = imageio.imread(os.path.join(compress_dir, f"{param_name}.png"))
+    img_norm = img / (2**8 - 1)
+
+    grid_norm = torch.tensor(img_norm)
+    mins = torch.tensor(meta["mins"])
+    maxs = torch.tensor(meta["maxs"])
+    grid = grid_norm * (maxs - mins) + mins
+
+    params = grid.reshape(meta["shape"])
+    params = params.to(dtype=getattr(torch, meta["dtype"]))
+    return params
+
+
+def _compress_png_16bit(
+    compress_dir: str, param_name: str, params: Tensor, n_sidelen: int, **kwargs
+) -> dict[str, Any]:
+    """Compress parameters with 16-bit quantization and PNG compression.
+
+    Args:
+        compress_dir (str): compression directory
+        param_name (str): parameter field name
+        params (Tensor): parameters
+        n_sidelen (int): image side length
+
+    Returns:
+        dict[str, Any]: metadata
+    """
+    grid = params.reshape((n_sidelen, n_sidelen, -1))
+    mins = torch.amin(grid, dim=(0, 1))
+    maxs = torch.amax(grid, dim=(0, 1))
+    grid_norm = (grid - mins) / (maxs - mins)
+    img_norm = grid_norm.detach().cpu().numpy()
+    img = (img_norm * (2**16 - 1)).round().astype(np.uint16)
+
+    img_l = img & 0xFF
+    img_u = (img >> 8) & 0xFF
+    imageio.imwrite(
+        os.path.join(compress_dir, f"{param_name}_l.png"), img_l.astype(np.uint8)
+    )
+    imageio.imwrite(
+        os.path.join(compress_dir, f"{param_name}_u.png"), img_u.astype(np.uint8)
+    )
+
+    meta = {
+        "shape": list(params.shape),
+        "dtype": str(params.dtype).split(".")[1],
+        "mins": mins.tolist(),
+        "maxs": maxs.tolist(),
+    }
+    return meta
+
+
+def _decompress_png_16bit(
+    compress_dir: str, param_name: str, meta: dict[str, Any]
+) -> Tensor:
+    """Decompress parameters from PNG files.
+
+    Args:
+        compress_dir (str): compression directory
+        param_name (str): parameter field name
+        meta (dict[str, Any]): metadata
+
+    Returns:
+        Tensor: parameters
+    """
+    img_l = imageio.imread(os.path.join(compress_dir, f"{param_name}_l.png"))
+    img_u = imageio.imread(os.path.join(compress_dir, f"{param_name}_u.png"))
+    img_u = img_u.astype(np.uint16)
+    img = (img_u << 8) + img_l
+
+    img_norm = img / (2**16 - 1)
+    grid_norm = torch.tensor(img_norm)
+    mins = torch.tensor(meta["mins"])
+    maxs = torch.tensor(meta["maxs"])
+    grid = grid_norm * (maxs - mins) + mins
+
+    params = grid.reshape(meta["shape"])
+    params = params.to(dtype=getattr(torch, meta["dtype"]))
+    return params
+
+
+def _compress_npz(compress_dir: str, param_name: str, params: Tensor) -> dict[str, Any]:
+    """Compress parameters with numpy's NPZ compression."""
+    npz_dict = {"arr": params.detach().cpu().numpy()}
+    np.savez_compressed(os.path.join(compress_dir, f"{param_name}.npz"), **npz_dict)
+    meta = {
+        "shape": params.shape,
+        "dtype": str(params.dtype).split(".")[1],
+    }
+    return meta
+
+
+def _decompress_npz(compress_dir: str, param_name: str, meta: dict[str, Any]) -> Tensor:
+    """Decompress parameters with numpy's NPZ compression."""
+    arr = np.load(os.path.join(compress_dir, f"{param_name}.npz"))["arr"]
+    params = torch.tensor(arr)
+    params = params.reshape(meta["shape"])
+    params = params.to(dtype=getattr(torch, meta["dtype"]))
+    return params
+
+
+def _compress_kmeans(
+    compress_dir: str,
+    param_name: str,
+    params: Tensor,
+    n_clusters: int = 65536,
+    verbose: bool = True,
+    **kwargs,
+) -> dict[str, Any]:
+    """Run K-means clustering on parameters and save centroids and labels to a npz file.
+
+    .. warning::
+        TorchPQ must installed to use K-means clustering.
+
+    Args:
+        compress_dir (str): compression directory
+        param_name (str): parameter field name
+        params (Tensor): parameters to compress
+        n_clusters (int): number of K-means clusters
+        verbose (bool, optional): Whether to print verbose information. Default to True.
+
+    Returns:
+        dict[str, Any]: metadata
+    """
+    try:
+        from torchpq.clustering import KMeans
+    except:
+        raise ImportError(
+            "Please install torchpq with 'pip install torchpq' to use K-means clustering"
+        )
+
+    kmeans = KMeans(n_clusters=n_clusters, distance="manhattan", verbose=verbose)
+    x = params.reshape(params.shape[0], -1).permute(1, 0).contiguous()
+    labels = kmeans.fit(x)
+    labels = labels.detach().cpu().numpy()
+    centroids = kmeans.centroids.permute(1, 0)
+
+    mins = torch.min(centroids)
+    maxs = torch.max(centroids)
+    centroids_norm = (centroids - mins) / (maxs - mins)
+    centroids_norm = centroids_norm.detach().cpu().numpy()
+    centroids_quant = (centroids_norm * (2**6 - 1)).round().astype(np.uint8)
+    labels = labels.astype(np.uint16)
+
+    npz_dict = {
+        "centroids": centroids_quant,
+        "labels": labels,
+    }
+    np.savez_compressed(os.path.join(compress_dir, f"{param_name}.npz"), **npz_dict)
+    meta = {
+        "shape": list(params.shape),
+        "dtype": str(params.dtype).split(".")[1],
+        "mins": mins.tolist(),
+        "maxs": maxs.tolist(),
+    }
+    return meta
+
+
+def _decompress_kmeans(
+    compress_dir: str, param_name: str, meta: dict[str, Any], **kwargs
+) -> Tensor:
+    """Decompress parameters from K-means compression.
+
+    Args:
+        compress_dir (str): compression directory
+        param_name (str): parameter field name
+        meta (dict[str, Any]): metadata
+
+    Returns:
+        Tensor: parameters
+    """
+    npz_dict = np.load(os.path.join(compress_dir, f"{param_name}.npz"))
+    centroids_quant = npz_dict["centroids"]
+    labels = npz_dict["labels"]
+
+    centroids_norm = centroids_quant / (2**6 - 1)
+    centroids_norm = torch.tensor(centroids_norm)
+    mins = torch.tensor(meta["mins"])
+    maxs = torch.tensor(meta["maxs"])
+    centroids = centroids_norm * (maxs - mins) + mins
+
+    params = centroids[labels]
+    params = params.reshape(meta["shape"])
+    params = params.to(dtype=getattr(torch, meta["dtype"]))
+    return params
